Make migrations resilient to app refactors

+++ apps/screens/migrations/0002_seed_lcd1602.py
@@ -1,11 +1,16 @@
 from django.db import migrations
 
+from apps.migration_utils import get_model
+
 
 LCD_SLUG = "lcd-1602"
 
 
 def seed_lcd_screen(apps, schema_editor):
-    DeviceScreen = apps.get_model("screens", "DeviceScreen")
+    DeviceScreen = get_model(apps, "screens", "DeviceScreen", allow_missing=True)
+
+    if DeviceScreen is None:
+        return
 
     defaults = {
         "name": "I2C LCD1602",
@@ -30,7 +35,11 @@ def seed_lcd_screen(apps, schema_editor):
 
 
 def remove_lcd_screen(apps, schema_editor):
-    DeviceScreen = apps.get_model("screens", "DeviceScreen")
+    DeviceScreen = get_model(apps, "screens", "DeviceScreen", allow_missing=True)
+
+    if DeviceScreen is None:
+        return
+
     DeviceScreen.all_objects.filter(slug=LCD_SLUG).delete()
